+++ packages/dataviews/src/dataform-layouts/regular/index.tsx
@@ -46,6 +46,7 @@ export default function FormRegularField< Item >( {
 	field,
 	onChange,
 	hideLabelFromVision,
+	validity,
 }: FieldLayoutProps< Item > ) {
 	const { fields } = useContext( DataFormContext );
 
@@ -67,6 +68,7 @@ export default function FormRegularField< Item >( {
 					data={ data }
 					form={ form }
 					onChange={ onChange }
+					validity={ validity?.children }
 				/>
 			</>
 		);
@@ -110,6 +112,7 @@ export default function FormRegularField< Item >( {
 							field={ fieldDefinition }
 							onChange={ onChange }
 							hideLabelFromVision
+							validity={ validity }
 						/>
 					) }
 				</div>
@@ -141,6 +144,7 @@ export default function FormRegularField< Item >( {
 					hideLabelFromVision={
 						labelPosition === 'none' ? true : hideLabelFromVision
 					}
+					validity={ validity }
 				/>
 			) }
 		</div>