+++ packages/editor/src/components/collab-sidebar/index.js
@@ -27,6 +27,7 @@ import {
 	useEnableFloatingSidebar,
 } from './hooks';
 import { focusCommentThread } from './utils';
+import PluginMoreMenuItem from '../plugin-more-menu-item';
 
 function CollabSidebarContent( {
 	showCommentBoard,
@@ -195,6 +196,9 @@ export default function CollabSidebar() {
 					/>
 				</PluginSidebar>
 			) }
+			<PluginMoreMenuItem icon={ commentIcon } onClick={ openTheSidebar }>
+				{ __( 'Notes' ) }
+			</PluginMoreMenuItem>
 		</>
 	);
 }